small code unifications
